Hotfix to give SCTP listener correct stream IDs
The eNodeB will not accept SCTP messages relating to the
S1SetupRequest if they have a stream ID != 0, and will not accept
UE SCTP messages if they have a stream ID != 1. This is a very quick
and dirty hotfix to patch that - taking advantage of the fact that
(in the current single eNB/UE setup), the first message sent by
the listener will be the S1Setup message, and all other messages
will be UE messages. (So stream ID is 0 for the first message, and
1 for all other messages).