With few exceptions, functions should return status_t
Change the code so that the majority of functions return status_t,
indicating whether the function succeeded or failed. This required
changes to the S1AP handling code, which previously returned whether
the function being handled has a response or not - this outcome is
now indicated through a parameter.

The few exceptions to this are (for example) code in libraries or
files that does not link to libcore, and therefore does not have
(and should not have) access to the status_t type.