Use libcore for uniform error handling
Replace the if/else, printf and other error handling code, and
replace it with the debug functions provided by NextEPC libcore,
in particular d_info (for printf equivalent) and d_assert for
error handling.

A consequence of this change is that log.h, which defined the
functions printOK, printWarning, printError and printInfo, is
no longer needed (and so has been removed).