Modified InitialContextSequpRequest to cope with new UE_NAS_SQN
The way that the DB handles UE_NAS_SQN has changed - it no longer
subtracts 1 from the value, so the InitialContextSetupRequest
must do this itself.

Note that this approach (getting UE_NAS_SQN from the DB then
extracting 1 from it) is not a suitable approach - it will cause
bugs later down the road, and this approach has been marked as
requiring a rewrite.