Refactor attach request to be separate from InitialUEMessage
Previously an initialUEMessage was treated as being
equivalent to an attach request. However, while the Attach Request
is delivered inside an InitialUEMessage, it is not the only message
to be delivered this way. The detach request (not yet implemented)
is one such message, and this refactoring is necessary work before
the detach request can be implemented.