Use the UplinkNASTransport variant of the Detach Request
The previous version of the Detach Request assumed it arrived
in an InitialUEMessage. This caused some problems, because the
NAS-PDU in the InitialUEMEssage needs to be deprotected, but
the InitialUEMessage does not contain a MME_UE_S1AP_ID and therefore
the DB cannot be accessed to get the deprotection keys.

However, Nervion encapsulates the detach request in the
UplinkNASTransport message, which *does* contain the InitialUEMessage,
and so the core has been changed to work with this variant of the
detach request.