Fix incorrect assumption when extracting DB values
I had incorrectly assumed that the DB provided the UE_NAS_SQN
as starting from 1, while I needed it to start from zero, so I
added code to subtract 1 every time it was retrieved.

However this assumption was false: it looked as the the SQN was
starting from one because it auto-increments, and I accidentally
retrieved it from the DB but never used it during the handling of
the attach request.

This has now been fixed.