Fix bug in handling NAS sequence numbers
There was a bug in handling NAS sequence numbers, whereby only the
sixth byte (the actual sequence number) was set. This was a
problem because it might be the case that the preceding 5 bytes
were not cleared, resulting in an incorrect sequence number.

Now the entire 16 bytes of that line are cleared before the sixth
byte is set.