Fix bug when creating threads
Fix a bug where created threads do not exit properly, resulting in
system resources being used up even once the thread has finished
executing. By default, threads are created as JOINABLE, and their
resources are not fully released until the thread has been joined.
Since the worker threads in corekube are never joined, this results
in system resources being used up for each thread created, until
no more resources are available and creating new threads fails with
the EAGAIN error.

The fix for this bug is to create threads as DETACHED, rather than
joinable, as this ensures that the resources used by a thread are
released when it terminates, without the need to join the thread.