Rework handling of responses
The previous system of handling responses was based upon an initial
system that assumed that each incoming message had a single outgoing
response. However (at least for 4G) it is possible for a single
incoming message to have more than one response, and so the response
system was 'bodged' to allow for a second response.

This 'bodge' was inelegant, and the response system has now been
reworked to allow for an arbitrary number of responses.