Romove incorrect call to ogs_free()
The call to ogs_free() was incorrect for two reasons:
  (a) the nasPdu structure in question had already been freed in
      the ngap_build_pdu_session_resource_setup_request() function
  (b) the structure is a packet buffer (ogs_pkbuf_t) and so
      should be freed with ogs_pkbuf_free(), rather than ogs_free()

The combination of these two factors produced a rather unusual bug.
While the code worked fine most of the time, occasionally it would
fail because the pointer to the memory pool used by libogscore would
be overwritten and therefore things would break (in particular,
the visible error would be that the mutex could not be obtained to
perform a memory operation).