Bugfix: Ensure every DB push has a pull to avoid errors in the DB protocol
There is an error in the DB protocol implementation whereby doing
a push without a corresponding pull can cause a later DB access to
fail when it receives null bytes.

This error may(?) be caused by the DB sending a reply to the push-only
access, even though no reply is expected. This reply, consisting
of only null bytes, is then read by the next DB access instead of
the values it was trying to access. This causes the error.

Rather than reimplementing the DB protocol, ensuring that every DB
push has a correspinding pull prevents this error from occuring, at
the slight cost of increased communication inefficiencies (transferring
an extra value from the DB that will not be used).