Bugfix: UE Context Release Command used wrong ID in message
The UE Context Release Command used the RAN ID when the AMF ID was
expected. In many situations (particularly testing / debug) these
IDs are equal, which prevented this bug from being detected until
now.

An additional fix was applied to the code: in most circumstances
the RAN/AMF ID keypair should be used, with the AMF ID being used
only when the RAN ID is unknown. Previously only the AMF ID was
sent; now it is sent only if the RAN ID is unknown.