Add support for SQN resync procedure (via Authentication Failure)
If the RAN detects that the SQN that CoreKube uses is out of sync,
it will send a Authentication Failure message. Previously CoreKube
simply crashed, now it correctly calculates the updated SQN and
re-attempts the authentication.

The default SQN that CoreKube uses has been set to null bytes; in
a future release this value will be stored in the DB for subsequent
uses.