Better subprocess interface with string vectors
Use explicit string vectors to define command instead of trying to
split a single string at spaces, because there were problems dealing
with arguments that contain spaces.

Also, this commit adds stream_cout as a boolean parameter to .run()
to allow the user to choose whether to print the output of the command
to the console or not.

Finally, the commit also adds exit code checking to the .run() method
to throw an exception if a command fails.