Disable running CloudFormation event query on multiple paginated responses
With --output=text, the query for AWS CLI is run on each page of a
paginated API call, returning the first match from each page. This
results in more than one match being returned when filtering for a
single event. We disable this behavior by using --no-paginate which
will return only the first page of results.