ci: Fix negative path filters not working since we can't change OR to AND
Negative path filters do not work currently since they are treated as
OR. We should be able to set predicate-quantifier: 'every' on the action
input to change it to AND, but due to a bug (dorny/paths-filter#225),
this does not work.

So we get rid of negative path filters and use only positive ones.